+++ packages/ack/lib/ack.dart
@@ -33,5 +33,6 @@ export 'src/schemas/schema.dart'
     hide AnyAckSchema, Refinement, SchemaOperation, WrapperSchema;
 export 'src/validation/ack_exception.dart';
 export 'src/validation/schema_error.dart';
+export 'package:standard_schema/standard_schema.dart';
 // Validation results
 export 'src/validation/schema_result.dart';